# BE/api-docs.yaml
openapi: 3.1.0
info:
  title: My API
  description: Descrizione dell'API
  version: "1.0"
servers:
- url: http://localhost:8080
  description: Generated server url
security:
- bearerAuth: []
paths:
  /api/public/insert-account:
    post:
      tags:
      - account-controller
      operationId: createAccount
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Account"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Account"
  /api/private/update-balance:
    post:
      tags:
      - card-controller
      operationId: updateCard
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Card"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: string
...
  /api/private/insert-transaction:
    post:
      tags:
      - transaction-controller
      operationId: insertTransaction
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Transaction"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Transaction"
...
components:
  schemas:
    Account:
      type: object
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string
        password:
          type: string
        user:
          $ref: "#/components/schemas/User"
    Card:
      type: object
      properties:
        id:
          type: integer
          format: int64
        type:
          type: string
          enum:
          - VIRTUAL
          - DEBIT
          - PREPAID
        balance:
          type: number
          format: double
        user:
          $ref: "#/components/schemas/User"
        finanziamenti:
          $ref: "#/components/schemas/Finanziamenti"
        transactions:
          type: array
          items:
            $ref: "#/components/schemas/Transaction"
    Finanziamenti:
      type: object
      properties:
        id:
          type: integer
          format: int64
        capitale:
          type: number
          format: double
        totRate:
          type: integer
          format: int32
        ratePagate:
          type: integer
          format: int32
        costoMensile:
          type: integer
          format: int32
        user:
          $ref: "#/components/schemas/User"
        card:
          $ref: "#/components/schemas/Card"
    Transaction:
      type: object
      properties:
        id:
          type: integer
          format: int64
        amount:
          type: number
          format: double
        description:
          type: string
        categoria:
          type: string
        timestamp:
          type: string
          format: date-time
        card:
          $ref: "#/components/schemas/Card"
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        email:
          type: string
        cognome:
          type: string
        iban:
          type: string
        cards:
          type: array
          items:
            $ref: "#/components/schemas/Card"
        finanziamenti:
          $ref: "#/components/schemas/Finanziamenti"
        account:
          $ref: "#/components/schemas/Account"
        random:
          type: object
          properties:
            seed:
              type: integer
              format: int64
              writeOnly: true
            deprecated:
              type: boolean
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT